refatora selected
